+++ crates/net/network/Cargo.toml
@@ -54,7 +54,7 @@ secp256k1 = { version = "0.24", features = [
 
 [dev-dependencies]
 # reth
-reth-discv4 = { path = "../discv4", features = ["mock"] }
+reth-discv4 = { path = "../discv4", features = ["test-utils"] }
 reth-interfaces = { path = "../../interfaces", features = ["test-utils"] }
 reth-provider = { path = "../../storage/provider", features = ["test-utils"] }
 reth-tracing = { path = "../../tracing" }
